feat: making some tests better
