+++ lib/movements/view/movements_screen.dart
@@ -1,19 +1,24 @@
 import 'package:flutter/material.dart';
+import 'package:hooks_riverpod/hooks_riverpod.dart';
 
+import '../../shared/controller/movement_provider.dart';
 import '../../shared/templates/bottom_navigation_bar_app.dart';
 import '../widgets/body_movements.dart';
 
-class MovementsScreen extends StatelessWidget {
+class MovementsScreen extends HookConsumerWidget {
   const MovementsScreen({super.key});
 
   static const route = '/movements';
 
   @override
-  Widget build(BuildContext context) {
-    return const SafeArea(
+  Widget build(BuildContext context, WidgetRef ref) {
+    var movements = ref.watch(movementProvider.state).state;
+    return SafeArea(
       child: Scaffold(
-        body: BodyMovements(),
-        bottomNavigationBar: BottomNavigationBarApp(index: 1),
+        body: BodyMovements(
+          movements: movements,
+        ),
+        bottomNavigationBar: const BottomNavigationBarApp(index: 1),
       ),
     );
   }
