+++ lib/movements/widgets/body_movements.dart
@@ -1,19 +1,18 @@
 import 'package:flutter/material.dart';
-import 'package:hooks_riverpod/hooks_riverpod.dart';
+
+import 'package:projeto_crypto/movements/view/model/movement_model.dart';
 import 'package:projeto_crypto/movements/widgets/container_movements.dart';
 import 'package:projeto_crypto/movements/widgets/list_view_movements.dart';
 
-import '../../shared/controller/movement_provider.dart';
-
-class BodyMovements extends HookConsumerWidget {
+class BodyMovements extends StatelessWidget {
+  final List<MovementModel> movements;
   const BodyMovements({
     Key? key,
+    required this.movements,
   }) : super(key: key);
 
   @override
-  Widget build(BuildContext context, WidgetRef ref) {
-    var movements = ref.watch(movementProvider.state).state;
-
+  Widget build(BuildContext context) {
     return Padding(
       padding: const EdgeInsets.only(left: 10),
       child: Column(
