+++ test/widgets_test/movements_screen_test.dart
@@ -1,6 +1,5 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:projeto_crypto/movements/view/model/movement_model.dart';
 import 'package:projeto_crypto/movements/view/movements_screen.dart';
 import 'package:projeto_crypto/movements/widgets/body_movements.dart';
 import 'package:projeto_crypto/movements/widgets/container_movements.dart';
@@ -8,13 +7,18 @@ import 'package:projeto_crypto/movements/widgets/list_tile_movements.dart';
 import 'package:projeto_crypto/movements/widgets/list_view_movements.dart';
 import 'package:projeto_crypto/shared/templates/bottom_navigation_bar_app.dart';
 
+import '../helpers/crypto_mock_data.dart';
 import '../helpers/setup_widget_tester.dart';
 
 void main() {
   group('Testing movemnts screen', () {
     testWidgets('Testing if the body has all it needs',
         (WidgetTester tester) async {
-      await loadPage(tester, const BodyMovements());
+      await loadPage(
+          tester,
+          BodyMovements(
+            movements: [movement],
+          ));
 
       final dividerFinder = find.byType(Divider);
       final columnFinder = find.byType(Column);
@@ -49,34 +53,30 @@ void main() {
 
       final textFinder = find.byType(Text);
       final containerFinder = find.byType(Container);
+      final containerMovementsTextFinder = find.byType(ContainerMovementsText);
 
       expect(containerFinder, findsOneWidget);
       expect(textFinder, findsOneWidget);
+      expect(containerMovementsTextFinder, findsOneWidget);
     });
 
     testWidgets('Testing ListView movements', (WidgetTester tester) async {
       await loadPage(
           tester,
           ListViewMovements(
-            movements: [
-              MovementModel(
-                  converted: '',
-                  received: '',
-                  date: DateTime.now(),
-                  valueInReal: '',
-                  discount: 0,
-                  increase: 0,
-                  idDiscount: '',
-                  idIncrease: '')
-            ],
+            movements: [movement],
           ));
       await tester.pumpAndSettle();
 
       final listViewFinder = find.byType(ListView);
       final listTileFinder = find.byType(ListTileMovements);
+      final ListTileMovements listTile = tester.widget(listTileFinder);
 
       expect(listViewFinder, findsOneWidget);
       expect(listTileFinder, findsOneWidget);
+      expect(listTile.convertQuantity, movement.converted);
+      expect(listTile.receiveQuantity, movement.received);
+      expect(listTile.total, movement.valueInReal);
     });
 
     testWidgets('Testing ListTile', (WidgetTester tester) async {
